update: rename potoke -> potoken
